requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==1.0.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
distro==1.9.0
Django==5.0.6
docx2txt==0.8
en-core-web-sm==2.3.1
exceptiongroup==1.2.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
llamaapi==0.1.36
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.33.0
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
plac==1.1.3
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==1.0.7
thinc==7.4.6
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wasabi==0.10.1
weasel==0.4.1
wrapt==1.16.0
yarl==1.9.4
